completed deleting messages correctly
