+++ src/javascripts/components/addMessages/addMessages.js
@@ -5,18 +5,11 @@ import data from '../../helpers/data/messages';
 import printer from '../placeholder/placeholder';
 import './addMessages.scss';
 
+
 let counter = 6;
 
-const addMessage = (e) => {
+const createNewMessage = (e) => {
   const typedValue = e.target.value;
-  // let newMessage = [];
-  // newMessage += `
-  //   <div class="newMessageContainer message">
-  //     <span class="closebtn">×</span>
-  //     <p>${typedValue}</p>
-  //     <div class="timestamp">${moment().format('LLL')}</div>
-  //   </div>`;
-  // $('#chatBox').prepend(newMessage);
   const newMessage = {
     name: 'Me',
     text: `${typedValue}`,
@@ -25,30 +18,15 @@ const addMessage = (e) => {
   };
   counter += 1;
   data.addToArray(newMessage);
-  printer.makeText();
-};
-
-const findObjectFromEvent = (e) => {
-  e.preventDefault();
-  const selection = e.target.id;
-  const messages = data.getMessages();
-  const messageIndex = messages.findIndex((message) => message.id === selection);
-  data.deleteFromArray(messageIndex);
-  console.log(messageIndex);
-  printer.makeText();
-};
-
-const deleteListener = () => {
-  $('.closebtn').on('click', findObjectFromEvent);
 };
 
 const inputCreateMessage = () => {
   $('#myInput').on('keyup', (e) => {
     if (e.keyCode === 13) {
       e.preventDefault();
-      addMessage(e);
+      createNewMessage(e);
       $('#myInput').val('');
-      deleteListener();
+      printer.messagePrinter();
     }
   });
 };
