+++ src/javascripts/components/placeholder/placeholder.js
@@ -1,24 +1,12 @@
 import utilities from '../../helpers/utilities';
 import './placeholder.scss';
-import m from '../../helpers/data/messages';
-import deleter from '../deleteMessages/deleteMessages';
+import data from '../../helpers/data/messages';
 
-const messages = m.getMessages();
+const messages = data.getMessages();
 
-const makeText = () => {
-  let domString = [];
+const messagePrinter = () => {
+  let domString = '';
   for (let i = 0; i < messages.length; i += 1) {
-  //   if (messages[i].name === 'Beetlejuice') {
-  //     domString += `
-  //   <div class="messageContainer message">
-  //     <span class="closebtn" id="${messages[i].id}">×</span>
-  //     <h6>${messages[i].name}</h6>
-  //     <p>${messages[i].text}</p>
-  //     <div class="timestamp">${messages[i].timeStamp}</div>
-  //   </div>
-  //   <br>
-  // `;
-  //   } else {
     domString += `
     <div class="messageContainer message" style="margin-right: 5px; margin-left: auto;">
       <span class="closebtn" id="${messages[i].id}">×</span>
@@ -26,12 +14,9 @@ const makeText = () => {
       <p>${messages[i].text}</p>
       <div class="timestamp">${messages[i].timeStamp}</div>
     </div>
-    <br>
   `;
   }
   utilities.printToDom('chatBox', domString);
-  deleter.deleteListener();
 };
 
-
-export default { makeText };
+export default { messagePrinter };
