+++ src/javascripts/main.js
@@ -7,14 +7,14 @@ import theme from './components/themeChanger/themeChange';
 import addMessages from './components/addMessages/addMessages';
 import largeText from './components/largetext/largeText';
 import customizer from './components/customizer/customizer';
+import deleter from './components/deleteMessages/deleteMessages';
 
 const init = () => {
   navbar.createNavbar();
-  placeholder.makeText();
+  placeholder.messagePrinter();
+  deleter.deleteListener();
   clearMessages.clearFunc();
-  // placeholder.timeStamp();
   theme.addCheckListener();
-  // addMessages.deleteMessage();
   addMessages.messageLimit();
   largeText.addLargeCheckListener();
   customizer.init();
