+++ src/styles/main.scss
@@ -39,6 +39,7 @@
       background-color: themed('messageBackground');
       color: themed('messageTextColor');
       box-shadow: themed('shadowBox');
+      margin: 10px;
     }
   }
   .newMessageContainer {
